[BETA] Preparation for foundryvtt 0.8.0
