+++ .gitignore
@@ -1,5 +1,2 @@
 node_modules
-foundryconfig.json
-.vscode
-releaseAnnouncement.txt
 /dist
