+++ changelog.md
@@ -1,23 +1,34 @@
+## [BETA] 0.2.3
+
+- Preparation for foundry 0.8.0
+- Bootstrapped with League of Extraordinary FoundryVTT Developers  [foundry-vtt-types](https://github.com/League-of-Foundry-Developers/foundry-vtt-types)
+- Added [Bug Reporter Module](https://foundryvtt.com/packages/bug-reporter/)
+- New Feature: Configurable option to use the player's color as blended background color of the message. by @Biggles
+- Translation: Updated German translation by @Biggles
+
 ## 0.2.2
 
-    - Integration with [Nick's East foundry-pc-types](https://gitlab.com/foundry-projects/foundry-pc/foundry-pc-types/-/tree/master)
-    - Some minor upgrade of the structure of the project
-    - [Fix: no token img if no token selected](https://github.com/ShoyuVanilla/FoundryVTT-Chat-Portrait/pull/22)
+- Integration with [Nick's East foundry-pc-types](https://gitlab.com/foundry-projects/foundry-pc/foundry-pc-types/-/tree/master)
+- Some minor upgrade of the structure of the project
+- [Fix: no token img if no token selected](https://github.com/ShoyuVanilla/FoundryVTT-Chat-Portrait/pull/22)
 
 ## 0.2.1
-    - Fix: Fixed the bug that the configuration window is not opened
-    - Translation: Spanish translation by @lozalojo
-    - Translation: German translation by @envy
+
+- Fix: Fixed the bug that the configuration window is not opened
+- Translation: Spanish translation by @lozalojo
+- Translation: German translation by @envy
 
 ## 0.2.0
-    - Migrated to TypeScript
-    - New Feature: Seperated configuration window
-    - New Feature: Configurable option for portrait border's color and width
-    - New Feature: Configurable option for placing chat flavor text next to portrait
-    - Fixed various issues
-    - Translation: Castilian translation by @HonzoNebro
+
+- Migrated to TypeScript
+- New Feature: Seperated configuration window
+- New Feature: Configurable option for portrait border's color and width
+- New Feature: Configurable option for placing chat flavor text next to portrait
+- Fixed various issues
+- Translation: Castilian translation by @HonzoNebro
 
 ## 0.1.4
-    - New Feature: Configurable option for using the player's color on chat message border. by @gatesvp
-    - New Feature: Configurable option for portrait image size
-    - Fix: Fixed the problem that non-square images are squeezed. by @megastruktur
+
+- New Feature: Configurable option for using the player's color on chat message border. by @gatesvp
+- New Feature: Configurable option for portrait image size
+- Fix: Fixed the problem that non-square images are squeezed. by @megastruktur
