+++ foundryconfig.json
@@ -1,0 +1,5 @@
+{
+  "dataPath": "D:/FoundryVTT",
+  "repository": "",
+  "rawURL": ""
+}
