--- makezip.sh
@@ -1,3 +1,0 @@
-cd ..
-zip -r itemcollection.zip itemcollection -x \*.git/* -x itemcollection/images/*  -x itemcollection/itemcollection.zip
-
