+++ package.json
@@ -1,31 +1,37 @@
 {
-	"name": "foundryvtt-chat-portrait",
-	"version": "0.2.1",
-	"description": "A Foundry VTT module that displays the Actor's portrait images on the chat messages.",
-	"main": "chat-portrait.js",
-	"author": "",
-	"license": "MIT",
-	"scripts": {
-		"package": "gulp package",
-		"build": "gulp build && gulp link",
-		"build:watch": "gulp watch",
-		"clean": "gulp clean && gulp link --clean",
-		"update": "npm install --save-dev gitlab:foundry-projects/foundry-pc/foundry-pc-types"
-	},
-	"devDependencies": {
-		"archiver": "^5.2.0",
-		"browser-sync": "^2.26.14",
-		"chalk": "^4.1.0",
-		"foundry-pc-types": "gitlab:foundry-projects/foundry-pc/foundry-pc-types",
-		"fs-extra": "^9.0.1",
-		"gulp": "^4.0.2",
-		"gulp-git": "^2.10.1",
-		"gulp-less": "^4.0.1",
-		"gulp-sass": "^4.1.0",
-		"gulp-typescript": "^6.0.0-alpha.1",
-		"json-stringify-pretty-compact": "^2.0.0",
-		"sass": "^1.26.10",
-		"typescript": "^4.0.2",
-		"yargs": "^15.4.1"
-	}
+  "private": true,
+  "name": "chat-portrait",
+  "version": "0.2.3",
+  "description": "Generate token bases based on token disposition or actor folder color.",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "package": "gulp package",
+    "build": "gulp build && gulp link",
+    "build:watch": "gulp watch",
+    "clean": "gulp clean && gulp link --clean",
+    "update": "npm install --save-dev @league-of-foundry-developers/foundry-vtt-types@fvtt-0.7.9"
+  },
+  "devDependencies": {
+    "@league-of-foundry-developers/foundry-vtt-types": "^0.7.9-6",
+    "archiver": "^5.2.0",
+    "chalk": "^4.1.0",
+    "fs-extra": "^9.0.1",
+    "gulp": "^4.0.2",
+    "gulp-git": "^2.10.1",
+    "gulp-less": "^4.0.1",
+    "gulp-sass": "^4.1.0",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "json-stringify-pretty-compact": "^2.0.0",
+    "sass": "^1.26.10",
+    "typescript": "^4.2.4",
+    "yargs": "^15.4.1"
+  },
+  "dependencies": {
+    "eslint": "^7.14.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^4.2.0"
+  }
 }
