+++ src/module/ChatPortraitSettings.ts
@@ -1,0 +1,11 @@
+export interface ChatPortraitSettings{
+    useTokenImage: boolean,
+    portraitSize: number,
+    borderShape: string,
+    useUserColorAsBorderColor: boolean,
+    borderColor: string,
+    borderWidth: number,
+    useUserColorAsChatBorderColor: boolean,
+    flavorNextToPortrait: boolean,
+    forceNameSearch: boolean
+}
