+++ tsconfig.json
@@ -1,18 +1,14 @@
 {
 	"compilerOptions": {
-		"moduleResolution": "Node",
+		"moduleResolution": "node",
 		"noEmit": false,
 		"strict": false,
-    "allowUmdGlobalAccess": true,
-		"target": "ES2020",
+		"target": "es2020",
 		"lib": [
 			"DOM",
 			"ES6",
 			"ES2020"
     	],
-		"types": ["foundry-pc-types"]
-	},
-  "exclude": [
-    "node_modules"
-  ]
+		"types": ["@league-of-foundry-developers/foundry-vtt-types"]
+	}
 }
